Extracted types to their own files.
