+++ src/types/base-touchable/BaseTouchable.tsx
@@ -1,0 +1,8 @@
+export type BaseTouchableProps = {
+  buttons: Array<BaseTouchableButtons>;
+};
+
+export type BaseTouchableButtons = {
+  text: string;
+  onPress: () => void;
+};
